Separating out constants
