+++ ChaosRecipeEnhancer.UI/BusinessLogic/Constants/FilterAssets.cs
@@ -1,0 +1,8 @@
+﻿namespace ChaosRecipeEnhancer.UI.BusinessLogic.Constants
+{
+    public static class FilterAssets
+    {
+        public const string DefaultNormalItemFilterStyleFilePath = @"Assets\FilterStyles\NormalItemsStyle.txt";
+        public const string DefaultInfluencedItemFilterStyleFilePath = @"Assets\FilterStyles\InfluencedItemsStyle.txt";
+    }
+}
