+++ ChaosRecipeEnhancer.UI/BusinessLogic/Constants/SoundAssets.cs
@@ -1,0 +1,10 @@
+﻿namespace ChaosRecipeEnhancer.UI.BusinessLogic.Constants
+{
+    public static class SoundAssets
+    {
+        public const string SoundFileExtensionFilter = "MP3|*.mp3";
+
+        public const string DefaultFilterChangedSoundFilePath = @"Assets\Sounds\FilterChanged.mp3";
+        public const string DefaultItemPickedUpSoundFilePath = @"Assets\Sounds\ItemPickedUp.mp3";
+    }
+}
