+++ ChaosRecipeEnhancer.UI/BusinessLogic/Constants/StringConstruction.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace ChaosRecipeEnhancer.UI.BusinessLogic.Constants
+{
+    public static class StringConstruction
+    {
+        public static string NewLineCharacter => Environment.NewLine;
+        public static string DoubleNewLineCharacter => Environment.NewLine + Environment.NewLine;
+        public const string TabCharacter = "\t";
+    }
+}
